pyproject.toml
[project]
name = "pyutils8ccr"
version = "0.22.0"
authors = [
  { name="9oOzv", email="miika.piiroinen@gmail.com" },
]
description = "My python utils package"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    'ruamel.yaml',
    'ruamel.yaml.string'
]

[project.urls]
Homepage = "https://github.com/9oOzv/pyutils8ccr"
Issues = "https://github.com/9oOzv/pyutils8ccr/issues"
